requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
